+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.5.0+1.21.3 - 2025-05-XX
+- More consistent clear input behavior in sync mode
+- Clear input also sets the mouse to a default position
+- MCIO_PROTOCOL_VERSION = 4
+
 ## 0.5.0+1.21.3 - 2025-04-06
 - Change protocol to combine keys and mouse buttons in the same list
 - MCIO_PROTOCOL_VERSION = 4